Put transactions module into single file
