--- src/store/modules/transactions/index.js
@@ -1,10 +1,0 @@
-import state from './state'
-import mutations from './mutations'
-import actions from './actions'
-
-export default {
-  namespaced: true,
-  state,
-  mutations,
-  actions
-}
