--- src/store/modules/transactions/state.js
@@ -1,3 +1,0 @@
-export default {
-  transactions: {}
-}
